coda adjustments

+++ src/index.ts
@@ -1,18 +1,22 @@
 import "dotenv/config";
-import express, { json } from "express";
 import "express-async-errors";
+import express, { json } from "express";
 import { petsRouter } from "./routes/petsRoutes";
+import { errorHandler } from "./middlewares/errorMidleware";
 import sequelize from "./config/database";
-import { errorHandler } from "./middlewares/errorMiddleware";
 
-const app = express();
+export const app = express();
 const port = 3000;
 
 app.use(json());
 app.use(petsRouter);
-app.use(errorHandler)
+app.use(errorHandler);
 
-sequelize.sync({ force: true }).then(() => {
-  console.log("Database is synced");
-  app.listen(port, () => console.log(`Listening at http://localhost:${port}`));
-});
+if (require.main === module) {
+  sequelize.sync({ force: true }).then(() => {
+    console.log("Database is synced");
+    app.listen(port, () =>
+      console.log(`Listening at http://localhost:${port}`),
+    );
+  });
+}
